+++ prospector/profiles/profile.py
@@ -109,6 +109,11 @@ def _is_valid_extension(filename):
 
 def _load_content(name_or_path, profile_path):
     filename = None
+    optional = False
+
+    if name_or_path.endswith("?"):
+        optional = True
+        name_or_path = name_or_path[:-1]
 
     if _is_valid_extension(name_or_path):
         for path in profile_path:
@@ -117,9 +122,6 @@ def _load_content(name_or_path, profile_path):
                 # this is a full path that we can load
                 filename = filepath
                 break
-
-        if filename is None:
-            raise ProfileNotFound(name_or_path, profile_path)
     else:
         for path in profile_path:
             for ext in ("yml", "yaml"):
@@ -128,8 +130,10 @@ def _load_content(name_or_path, profile_path):
                     filename = filepath
                     break
 
-        if filename is None:
-            raise ProfileNotFound(name_or_path, profile_path)
+    if filename is None:
+        if optional:
+            return {}
+        raise ProfileNotFound(name_or_path, profile_path)
 
     with open(filename) as fct:
         try: